+++ src/converters/_tunings_converters.ts
@@ -96,6 +96,14 @@ export function tuningDatasetToMldev(
     throw new Error('gcsUri parameter is not supported in Gemini API.');
   }
 
+  if (
+    common.getValueByPath(fromObject, ['vertexDatasetResource']) !== undefined
+  ) {
+    throw new Error(
+      'vertexDatasetResource parameter is not supported in Gemini API.',
+    );
+  }
+
   const fromExamples = common.getValueByPath(fromObject, ['examples']);
   if (fromExamples != null) {
     let transformedList = fromExamples;
@@ -105,18 +113,6 @@ export function tuningDatasetToMldev(
       });
     }
     common.setValueByPath(toObject, ['examples', 'examples'], transformedList);
-  }
-
-  return toObject;
-}
-
-export function tuningValidationDatasetToMldev(
-  fromObject: types.TuningValidationDataset,
-): Record<string, unknown> {
-  const toObject: Record<string, unknown> = {};
-
-  if (common.getValueByPath(fromObject, ['gcsUri']) !== undefined) {
-    throw new Error('gcsUri parameter is not supported in Gemini API.');
   }
 
   return toObject;
@@ -295,22 +291,6 @@ export function listTuningJobsParametersToVertex(
   return toObject;
 }
 
-export function tuningExampleToVertex(
-  fromObject: types.TuningExample,
-): Record<string, unknown> {
-  const toObject: Record<string, unknown> = {};
-
-  if (common.getValueByPath(fromObject, ['textInput']) !== undefined) {
-    throw new Error('textInput parameter is not supported in Vertex AI.');
-  }
-
-  if (common.getValueByPath(fromObject, ['output']) !== undefined) {
-    throw new Error('output parameter is not supported in Vertex AI.');
-  }
-
-  return toObject;
-}
-
 export function tuningDatasetToVertex(
   fromObject: types.TuningDataset,
   parentObject: Record<string, unknown>,
@@ -326,6 +306,17 @@ export function tuningDatasetToVertex(
     );
   }
 
+  const fromVertexDatasetResource = common.getValueByPath(fromObject, [
+    'vertexDatasetResource',
+  ]);
+  if (parentObject !== undefined && fromVertexDatasetResource != null) {
+    common.setValueByPath(
+      parentObject,
+      ['supervisedTuningSpec', 'trainingDatasetUri'],
+      fromVertexDatasetResource,
+    );
+  }
+
   if (common.getValueByPath(fromObject, ['examples']) !== undefined) {
     throw new Error('examples parameter is not supported in Vertex AI.');
   }
@@ -335,12 +326,24 @@ export function tuningDatasetToVertex(
 
 export function tuningValidationDatasetToVertex(
   fromObject: types.TuningValidationDataset,
+  parentObject: Record<string, unknown>,
 ): Record<string, unknown> {
   const toObject: Record<string, unknown> = {};
 
   const fromGcsUri = common.getValueByPath(fromObject, ['gcsUri']);
   if (fromGcsUri != null) {
     common.setValueByPath(toObject, ['validationDatasetUri'], fromGcsUri);
+  }
+
+  const fromVertexDatasetResource = common.getValueByPath(fromObject, [
+    'vertexDatasetResource',
+  ]);
+  if (parentObject !== undefined && fromVertexDatasetResource != null) {
+    common.setValueByPath(
+      parentObject,
+      ['supervisedTuningSpec', 'trainingDatasetUri'],
+      fromVertexDatasetResource,
+    );
   }
 
   return toObject;
@@ -359,7 +362,7 @@ export function createTuningJobConfigToVertex(
     common.setValueByPath(
       parentObject,
       ['supervisedTuningSpec'],
-      tuningValidationDatasetToVertex(fromValidationDataset),
+      tuningValidationDatasetToVertex(fromValidationDataset, toObject),
     );
   }
 
@@ -463,12 +466,6 @@ export function createTuningJobParametersToVertex(
   return toObject;
 }
 
-export function tunedModelCheckpointFromMldev(): Record<string, unknown> {
-  const toObject: Record<string, unknown> = {};
-
-  return toObject;
-}
-
 export function tunedModelFromMldev(
   fromObject: types.TunedModel,
 ): Record<string, unknown> {
@@ -567,6 +564,16 @@ export function tuningJobFromMldev(
   const fromPipelineJob = common.getValueByPath(fromObject, ['pipelineJob']);
   if (fromPipelineJob != null) {
     common.setValueByPath(toObject, ['pipelineJob'], fromPipelineJob);
+  }
+
+  const fromSatisfiesPzi = common.getValueByPath(fromObject, ['satisfiesPzi']);
+  if (fromSatisfiesPzi != null) {
+    common.setValueByPath(toObject, ['satisfiesPzi'], fromSatisfiesPzi);
+  }
+
+  const fromSatisfiesPzs = common.getValueByPath(fromObject, ['satisfiesPzs']);
+  if (fromSatisfiesPzs != null) {
+    common.setValueByPath(toObject, ['satisfiesPzs'], fromSatisfiesPzs);
   }
 
   const fromServiceAccount = common.getValueByPath(fromObject, [
@@ -816,6 +823,16 @@ export function tuningJobFromVertex(
   const fromPipelineJob = common.getValueByPath(fromObject, ['pipelineJob']);
   if (fromPipelineJob != null) {
     common.setValueByPath(toObject, ['pipelineJob'], fromPipelineJob);
+  }
+
+  const fromSatisfiesPzi = common.getValueByPath(fromObject, ['satisfiesPzi']);
+  if (fromSatisfiesPzi != null) {
+    common.setValueByPath(toObject, ['satisfiesPzi'], fromSatisfiesPzi);
+  }
+
+  const fromSatisfiesPzs = common.getValueByPath(fromObject, ['satisfiesPzs']);
+  if (fromSatisfiesPzs != null) {
+    common.setValueByPath(toObject, ['satisfiesPzs'], fromSatisfiesPzs);
   }
 
   const fromServiceAccount = common.getValueByPath(fromObject, [