+++ test/system/node/client_test.ts
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {fail} from 'assert';
 import {GoogleAuthOptions} from 'google-auth-library';
 import {z} from 'zod';
 import {zodToJsonSchema} from 'zod-to-json-schema';
@@ -168,7 +167,7 @@ describe('Client Tests', () => {
         }
       }
     });
-    it('ML Dev should generate content with given valid json schema', async () => {
+    it('ML Dev should generate content with given valid json schema in responseSchema', async () => {
       const innerObject = z.object({
         innerString: z.string(),
         innerNumber: z.number(),
@@ -211,7 +210,7 @@ describe('Client Tests', () => {
       expect(validationResult.success).toEqual(true);
     });
 
-    it('Vertex AI should generate content with given valid json schema', async () => {
+    it('Vertex AI should generate content with given valid json schema in responseSchema', async () => {
       const innerObject = z.object({
         innerString: z.string(),
         innerNumber: z.number(),
@@ -304,7 +303,184 @@ describe('Client Tests', () => {
       });
     });
   });
-  it('Vertex AI should be able to help build the FunctionDeclaration', async () => {
+  it('ML Dev should generate content with given valid json schema in responseJsonSchema', async () => {
+    const innerObject = z.object({
+      innerString: z.string(),
+      innerNumber: z.number(),
+    });
+    const nullableInnerObject = z.object({
+      innerString: z.string(),
+      innerNumber: z.number(),
+    });
+    const nestedSchema = z.object({
+      simpleString: z.string().describe('This is a simple string'),
+      stringDatatime: z.string().datetime(),
+      stringWithEnum: z.enum(['enumvalue1', 'enumvalue2', 'enumvalue3']),
+      stringWithLength: z.string().min(1).max(10),
+      simpleNumber: z.number(),
+      simpleInteger: z.number().int(),
+      integerInt64: z.number().int(),
+      numberWithMinMax: z.number().min(1).max(10),
+      simpleBoolean: z.boolean(),
+      arrayFiled: z.array(z.string()),
+      unionField: z.union([z.string(), z.number()]),
+      nullableField: z.string().nullable(),
+      nullableArrayField: z.array(z.string()).nullable(),
+      nullableObjectField: nullableInnerObject.nullable(),
+      inner: innerObject,
+    });
+    const client = new GoogleGenAI({vertexai: false, apiKey: GOOGLE_API_KEY});
+    const response = await client.models.generateContent({
+      model: 'gemini-1.5-flash',
+      contents: 'populate the following object',
+      config: {
+        responseMimeType: 'application/json',
+        responseJsonSchema: zodToJsonSchema(nestedSchema),
+      },
+    });
+    const parsedResponse = JSON.parse(
+      response.candidates![0].content!['parts']![0].text as string,
+    );
+    console.log('mldev response', parsedResponse);
+    const validationResult = nestedSchema.safeParse(parsedResponse);
+    expect(validationResult.success).toEqual(true);
+  });
+
+  it('Vertex AI should generate content with given valid json schema in responseJsonSchema', async () => {
+    const innerObject = z.object({
+      innerString: z.string(),
+      innerNumber: z.number(),
+    });
+    const nullableInnerObject = z.object({
+      innerString: z.string(),
+      innerNumber: z.number(),
+    });
+    const nestedSchema = z.object({
+      simpleString: z.string().default('default'),
+      stringDatetime: z.string().datetime(),
+      stringWithEnum: z.enum(['enumvalue1', 'enumvalue2', 'enumvalue3']),
+      stringWithLength: z.string().min(1).max(10),
+      simpleNumber: z.number(),
+      simpleInteger: z.number().int(),
+      integerInt64: z.number().int(),
+      numberWithMinMax: z.number().min(1).max(10),
+      simpleBoolean: z.boolean(),
+      arrayFiled: z.array(z.string()),
+      unionField: z.union([z.string(), z.number()]),
+      nullableField: z.string().nullable(),
+      nullableArrayField: z.array(z.string()).nullable(),
+      nullableObjectField: nullableInnerObject.nullable(),
+      inner: innerObject,
+    });
+
+    const client = new GoogleGenAI({
+      vertexai: true,
+      project: GOOGLE_CLOUD_PROJECT,
+      location: GOOGLE_CLOUD_LOCATION,
+    });
+    const response = await client.models.generateContent({
+      model: 'gemini-1.5-flash',
+      contents: 'populate the following object',
+      config: {
+        responseMimeType: 'application/json',
+        responseJsonSchema: zodToJsonSchema(nestedSchema),
+      },
+    });
+    const parsedResponse = JSON.parse(
+      response.candidates![0].content!['parts']![0].text as string,
+    );
+    console.log('vertex ai response', parsedResponse);
+    const validationResult = nestedSchema.safeParse(parsedResponse);
+    expect(validationResult.success).toEqual(true);
+  });
+  it('ML Dev can use JSON schema in parameters to build FunctionDeclaration', async () => {
+    const stringArgument = z.object({
+      firstString: z.string(),
+      secondString: z.string(),
+    });
+
+    const client = new GoogleGenAI({vertexai: false, apiKey: GOOGLE_API_KEY});
+    const response = await client.models.generateContent({
+      model: 'gemini-1.5-flash',
+      contents: 'put word: hello and word: world into a string',
+      config: {
+        tools: [
+          {
+            functionDeclarations: [
+              {
+                name: 'concatStringFunction',
+                description: 'this is a concat string function',
+                parameters: zodToJsonSchema(stringArgument) as Record<
+                  string,
+                  unknown
+                >,
+              },
+            ],
+          },
+        ],
+        toolConfig: {
+          functionCallingConfig: {
+            mode: FunctionCallingConfigMode.ANY,
+            allowedFunctionNames: ['concatStringFunction'],
+          },
+        },
+      },
+    });
+    const functionCallResponse =
+      response.candidates![0].content!['parts']![0].functionCall;
+    expect(functionCallResponse!.name).toEqual('concatStringFunction');
+    const parsedArgument = stringArgument.safeParse(
+      functionCallResponse!.args!,
+    );
+    expect(parsedArgument.success).toEqual(true);
+    expect(parsedArgument.data).toEqual({
+      firstString: 'hello',
+      secondString: 'world',
+    });
+  });
+  it('ML Dev can use JSON schema in parametersJsonSchema to build FunctionDeclaration', async () => {
+    const stringArgument = z.object({
+      firstString: z.string(),
+      secondString: z.string(),
+    });
+
+    const client = new GoogleGenAI({vertexai: false, apiKey: GOOGLE_API_KEY});
+    const response = await client.models.generateContent({
+      model: 'gemini-1.5-flash',
+      contents: 'put word: hello and word: world into a string',
+      config: {
+        tools: [
+          {
+            functionDeclarations: [
+              {
+                name: 'concatStringFunction',
+                description: 'this is a concat string function',
+                parametersJsonSchema: zodToJsonSchema(stringArgument),
+              },
+            ],
+          },
+        ],
+        toolConfig: {
+          functionCallingConfig: {
+            mode: FunctionCallingConfigMode.ANY,
+            allowedFunctionNames: ['concatStringFunction'],
+          },
+        },
+      },
+    });
+    const functionCallResponse =
+      response.candidates![0].content!['parts']![0].functionCall;
+    expect(functionCallResponse!.name).toEqual('concatStringFunction');
+    const parsedArgument = stringArgument.safeParse(
+      functionCallResponse!.args!,
+    );
+    expect(parsedArgument.success).toEqual(true);
+    expect(parsedArgument.data).toEqual({
+      firstString: 'hello',
+      secondString: 'world',
+    });
+  });
+  it('Vertex AI can use JSON schema in parameters to build FunctionDeclaration', async () => {
     const stringArgument = z.object({
       firstString: z.string(),
       secondString: z.string(),
@@ -328,6 +504,51 @@ describe('Client Tests', () => {
                   string,
                   unknown
                 >,
+              },
+            ],
+          },
+        ],
+        toolConfig: {
+          functionCallingConfig: {
+            mode: FunctionCallingConfigMode.ANY,
+            allowedFunctionNames: ['concatStringFunction'],
+          },
+        },
+      },
+    });
+    const functionCallResponse =
+      response.candidates![0].content!['parts']![0].functionCall;
+    expect(functionCallResponse!.name).toEqual('concatStringFunction');
+    const parsedArgument = stringArgument.safeParse(
+      functionCallResponse!.args!,
+    );
+    expect(parsedArgument.success).toEqual(true);
+    expect(parsedArgument.data).toEqual({
+      firstString: 'hello',
+      secondString: 'world',
+    });
+  });
+  it('Vertex AI can use JSON schema in parametersJsonSchema to build FunctionDeclaration', async () => {
+    const stringArgument = z.object({
+      firstString: z.string(),
+      secondString: z.string(),
+    });
+    const client = new GoogleGenAI({
+      vertexai: true,
+      project: GOOGLE_CLOUD_PROJECT,
+      location: GOOGLE_CLOUD_LOCATION,
+    });
+    const response = await client.models.generateContent({
+      model: 'gemini-1.5-flash',
+      contents: 'put word: hello and word: world into a string',
+      config: {
+        tools: [
+          {
+            functionDeclarations: [
+              {
+                name: 'concatStringFunction',
+                description: 'this is a concat string function',
+                parametersJsonSchema: zodToJsonSchema(stringArgument),
               },
             ],
           },
@@ -557,79 +778,6 @@ describe('Client Tests', () => {
 
     afterEach(function () {
       jasmine.DEFAULT_TIMEOUT_INTERVAL = originalTimeout;
-    });
-  });
-
-  describe('test async performance', () => {
-    it('generate content should complete in less than 1 second', async () => {
-      const client = new GoogleGenAI({vertexai: false, apiKey: GOOGLE_API_KEY});
-      async function firstAsyncFunc() {
-        client.models.generateContent({
-          model: 'gemini-1.5-flash',
-          contents: 'high',
-          config: {
-            systemInstruction: 'I say high you say low.',
-          },
-        });
-      }
-      async function secondAsyncFunc() {
-        client.models.generateContent({
-          model: 'gemini-1.5-flash',
-          contents: 'high',
-          config: {
-            systemInstruction: 'I say high you say low.',
-          },
-        });
-      }
-      const startTime = performance.now(); // Record start time
-      try {
-        await Promise.all([firstAsyncFunc(), secondAsyncFunc()]);
-      } catch (e) {
-        fail('Test failed due to error: ' + e);
-      } finally {
-        const endTime = performance.now(); // Record end time
-        const timeDelta = endTime - startTime;
-        expect(timeDelta).toBeLessThanOrEqual(
-          1050,
-          'Expected timeDelta to be less than or equal to 1050, got ' +
-            timeDelta,
-        );
-      }
-    });
-    it('stream generate content should complete in less than 1 second', async () => {
-      const client = new GoogleGenAI({vertexai: false, apiKey: GOOGLE_API_KEY});
-      async function firstAsyncFunc() {
-        client.models.generateContentStream({
-          model: 'gemini-1.5-flash',
-          contents: 'high',
-          config: {
-            systemInstruction: 'I say high you say low.',
-          },
-        });
-      }
-      async function secondAsyncFunc() {
-        client.models.generateContentStream({
-          model: 'gemini-1.5-flash',
-          contents: 'high',
-          config: {
-            systemInstruction: 'I say high you say low.',
-          },
-        });
-      }
-      const startTime = performance.now(); // Record start time
-      try {
-        await Promise.all([firstAsyncFunc(), secondAsyncFunc()]);
-      } catch (e) {
-        fail('Test failed due to error: ' + e);
-      } finally {
-        const endTime = performance.now(); // Record end time
-        const timeDelta = endTime - startTime;
-        expect(timeDelta).toBeLessThanOrEqual(
-          1050,
-          'Expected timeDelta to be less than or equal to 1050, got ' +
-            timeDelta,
-        );
-      }
     });
   });