+++ test/unit/mcp_test.ts
@@ -13,6 +13,7 @@ import {
 import {
   hasMcpToolUsage,
   mcpToTool,
+  setMcpToolUsageFromMcpToTool,
   setMcpUsageHeader,
 } from '../../src/mcp/_mcp.js';
 import * as types from '../../src/types.js';
@@ -20,188 +21,6 @@ import * as types from '../../src/types.js';
 import {spinUpPrintingServer} from './test_mcp_server.js';
 
 describe('mcpToGeminiTool', () => {
-  it('should filter unknown JsonSchema fields', () => {
-    const mcpTool = {
-      name: 'tool',
-      description: 'tool-description',
-      inputSchema: {
-        type: 'object',
-        properties: {},
-        unknownField: 'unknownField',
-        unknownObject: {},
-      },
-    };
-
-    const parsedMcpTool = ToolSchema.parse(mcpTool);
-    expect(mcpToGeminiTool(parsedMcpTool)).toEqual({
-      functionDeclarations: [
-        {
-          name: 'tool',
-          description: 'tool-description',
-          parameters: {
-            type: types.Type.OBJECT,
-            properties: {},
-          },
-        },
-      ],
-    });
-  });
-
-  it('should process items', () => {
-    const mcpTool = {
-      name: 'tool',
-      description: 'tool-description',
-      inputSchema: {
-        type: 'object',
-        properties: {
-          property: {
-            type: 'object',
-            items: {
-              type: 'string',
-              description: 'item-description',
-              unknownField: 'unknownField',
-              unknownObject: {},
-            },
-          },
-        },
-      },
-    };
-
-    const parsedMcpTool = ToolSchema.parse(mcpTool);
-    expect(mcpToGeminiTool(parsedMcpTool)).toEqual({
-      functionDeclarations: [
-        {
-          name: 'tool',
-          description: 'tool-description',
-          parameters: {
-            type: types.Type.OBJECT,
-            properties: {
-              property: {
-                type: types.Type.OBJECT,
-                items: {
-                  type: types.Type.STRING,
-                  description: 'item-description',
-                },
-              },
-            },
-          },
-        },
-      ],
-    });
-  });
-
-  it('should process anyOf', () => {
-    const mcpTool = {
-      name: 'tool',
-      description: 'tool-description',
-      inputSchema: {
-        type: 'object',
-        properties: {
-          property: {
-            anyOf: [
-              {
-                type: 'string',
-                description: 'anyOf-description-1',
-                unknownField: 'unknownField',
-                unknownObject: {},
-              },
-              {
-                type: 'number',
-                description: 'anyOf-description-2',
-                unknownField: 'unknownField',
-                unknownObject: {},
-              },
-            ],
-          },
-        },
-      },
-    };
-
-    const parsedMcpTool = ToolSchema.parse(mcpTool);
-    expect(mcpToGeminiTool(parsedMcpTool)).toEqual({
-      functionDeclarations: [
-        {
-          name: 'tool',
-          description: 'tool-description',
-          parameters: {
-            type: types.Type.OBJECT,
-            properties: {
-              property: {
-                anyOf: [
-                  {
-                    type: types.Type.STRING,
-                    description: 'anyOf-description-1',
-                  },
-                  {
-                    type: types.Type.NUMBER,
-                    description: 'anyOf-description-2',
-                  },
-                ],
-              },
-            },
-          },
-        },
-      ],
-    });
-  });
-
-  it('should process properties', () => {
-    const mcpTool = {
-      name: 'tool',
-      description: 'tool-description',
-      inputSchema: {
-        type: 'object',
-        properties: {
-          property: {
-            type: 'object',
-            properties: {
-              parameter1: {
-                type: 'string',
-                description: 'parameter1-description',
-                unknownField: 'unknownField',
-                unknownObject: {},
-              },
-              parameter2: {
-                type: 'string',
-                description: 'parameter2-description',
-                unknownField: 'unknownField',
-                unknownObject: {},
-              },
-            },
-          },
-        },
-      },
-    };
-
-    const parsedMcpTool = ToolSchema.parse(mcpTool);
-    expect(mcpToGeminiTool(parsedMcpTool)).toEqual({
-      functionDeclarations: [
-        {
-          name: 'tool',
-          description: 'tool-description',
-          parameters: {
-            type: types.Type.OBJECT,
-            properties: {
-              property: {
-                type: types.Type.OBJECT,
-                properties: {
-                  parameter1: {
-                    type: types.Type.STRING,
-                    description: 'parameter1-description',
-                  },
-                  parameter2: {
-                    type: types.Type.STRING,
-                    description: 'parameter2-description',
-                  },
-                },
-              },
-            },
-          },
-        },
-      ],
-    });
-  });
-
   it('should set behavior', () => {
     const mcpTool = {
       name: 'tool',
@@ -222,8 +41,8 @@ describe('mcpToGeminiTool', () => {
         {
           name: 'tool',
           description: 'tool-description',
-          parameters: {
-            type: types.Type.OBJECT,
+          parametersJsonSchema: {
+            type: 'object',
             properties: {},
           },
           behavior: types.Behavior.NON_BLOCKING,
@@ -279,13 +98,13 @@ describe('mcpToolsToGeminiTool', () => {
         {
           name: 'tool-1',
           description: 'tool-1-description',
-          parameters: {
-            type: types.Type.OBJECT,
+          parametersJsonSchema: {
+            type: 'object',
             properties: {
               property: {
-                type: types.Type.OBJECT,
+                type: 'object',
                 items: {
-                  type: types.Type.STRING,
+                  type: 'string',
                   description: 'item-description',
                 },
               },
@@ -295,13 +114,13 @@ describe('mcpToolsToGeminiTool', () => {
         {
           name: 'tool-2',
           description: 'tool-2-description',
-          parameters: {
-            type: types.Type.OBJECT,
+          parametersJsonSchema: {
+            type: 'object',
             properties: {
               property: {
-                type: types.Type.OBJECT,
+                type: 'object',
                 items: {
-                  type: types.Type.STRING,
+                  type: 'string',
                   description: 'item-description',
                 },
               },
@@ -324,13 +143,13 @@ describe('mcpToolsToGeminiTool', () => {
         {
           name: 'tool-1',
           description: 'tool-1-description',
-          parameters: {
-            type: types.Type.OBJECT,
+          parametersJsonSchema: {
+            type: 'object',
             properties: {
               property: {
-                type: types.Type.OBJECT,
+                type: 'object',
                 items: {
-                  type: types.Type.STRING,
+                  type: 'string',
                   description: 'item-description',
                 },
               },
@@ -341,13 +160,13 @@ describe('mcpToolsToGeminiTool', () => {
         {
           name: 'tool-2',
           description: 'tool-2-description',
-          parameters: {
-            type: types.Type.OBJECT,
+          parametersJsonSchema: {
+            type: 'object',
             properties: {
               property: {
-                type: types.Type.OBJECT,
+                type: 'object',
                 items: {
-                  type: types.Type.STRING,
+                  type: 'string',
                   description: 'item-description',
                 },
               },
@@ -368,6 +187,38 @@ describe('hasMcpToolUsage', () => {
   });
 
   it('should return false for Gemini tools', () => {
+    // Explicitly set the MCP usage to false.
+    setMcpToolUsageFromMcpToTool(false);
+    const tools = [
+      {
+        functionDeclarations: [
+          {
+            name: 'tool',
+            description: 'tool-description',
+            parametersJsonSchema: {
+              type: 'object',
+              properties: {
+                property: {
+                  type: 'object',
+                  items: {
+                    type: 'string',
+                    description: 'item-description',
+                  },
+                },
+              },
+            },
+          },
+        ],
+      },
+    ];
+    expect(hasMcpToolUsage(tools)).toBeFalse();
+  });
+
+  it('should return true after calling mcpToTool', async () => {
+    // Explicitly set the MCP usage to false.
+    setMcpToolUsageFromMcpToTool(false);
+
+    const _mcpCallableTool = mcpToTool(await spinUpPrintingServer());
     const tools = [
       {
         functionDeclarations: [
@@ -390,7 +241,7 @@ describe('hasMcpToolUsage', () => {
         ],
       },
     ];
-    expect(hasMcpToolUsage(tools)).toBeFalse();
+    expect(hasMcpToolUsage(tools)).toBeTrue();
   });
 });